+++ src/shared/scripts/content-scripts/parental-control.js
@@ -1,4 +1,8 @@
 (() => {
+  const baseUrl = 'https://raw.githubusercontent.com/roskomsvoboda/' +
+    'censortracker/gh-pages/static/img/parental-control/'
+  const randomInteger = () => Math.floor(Math.random() * 5) + 1
+
   const port = chrome.runtime.connect({ name: 'censortracker' })
 
   // Ask if parental control is isEnabled.
@@ -6,6 +10,8 @@
   port.onMessage.addListener((message) => {
     if (message.parentalControl) {
       const overlay = document.createElement('div')
+
+      overlay.style.backgroundImage = `url(${baseUrl}${randomInteger(1, 5)}.svg)`
 
       overlay.id = 'overlay'